Cleaning up tests after moving DatabaseInterface
